Remove ACPI fan and thermal zone support
Yes, this seems odd for a laptop, but we're removing ACPI fan and
thermal zone support. With thermal zones on, the fan spins up and
down constantly. Without it on, there's a more ordinary,
predictable fan pattern.

We're also disabling ACPI fan support, as there are no user-mode
programs which are controlling it. We'll let the platform
firmware do its job.