Presario_CQ57.config: No supported IOMMU; thermal back on
Thermal code is pulled in by the radeon driver, so we'll include
it. I've found no issues passing thermal.off=1 on the kernel
command line; your mileage may vary.

There's also no firmware support for the IOMMU and I don't think
this laptop has an old AGP GART, so we have that disabled.