Presario_CQ57.config: PS/2 IO re-enabled, experimental options
It seems the FCH on this computer handles the keyboard and mouse as
PS/2 devices instead of I2C, so those are back on and working.

We also turned on IOMMU support, as the datasheet for the E300
tells us the corresponding FCH supports it even though there's no
ACPI table.

We've also enabled a few FCH options and X2APIC support. There are
likely more. I need to start creating branches for experimenting
with individual options.